pyproject.toml
[project]
name = "docinfo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "argon2-cffi>=25.1.0",
    "fastapi>=0.124.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
